+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-12](https://github.com/ITK-Leantime/leantime-tickettemplate/pull/12)
+  * Migrate repository data access from raw PDO to the query builder.
+* [PR-11](https://github.com/ITK-Leantime/leantime-tickettemplate/pull/11)
+  * Normalize scaffolding.
 * [PR-10](https://github.com/ITK-Leantime/leantime-tickettemplate/pull/10)
   * Upgrade ticket template plugin leantime 3 4 3
   * Change markdown lint to use itkdev image